m4-evaluacion-credito 3

+++ m3-solicitud-credito/src/main/java/com/tingeso/m3_solicitud_credito/entity/FinEvalEntity.java
@@ -31,7 +31,7 @@ public class FinEvalEntity {
     private Integer currentJobAntiquity;
     private Boolean isSelfEmployed;
     private Boolean hasGoodIncomeHistory;
-    private Long monthlyDebt;
+    private Double monthlyDebt;
     //Para P4.R7.R74
     private Long bankAccountBalance;
     //Para P4.R7.R72
